requirements.txt
bcrypt==3.1.7
beautifulsoup4==4.9.3
cryptography==3.2.1
Django==3.1.3
django-mysql==3.9.0
django-sslserver==0.22
mysql-connector==2.2.9
pandas==1.1.4
XlsxWriter==1.1.2
requests
lxml
mysql-connector-python
